Work in progress; lots of changes
Consolidated everything into Model package

Added DataReader Interface that is extended by ReadFileData and ReadGUIData.

ReadGUIData is implemented to read a String from a JTextfield. ReadFileData is not yet implemented.

CalculateAge Class added to process the User input Date of Birth into the User's age, and handle a Birthday message (which still needs work).

Model Class created and partially implemented.

Login and LoginRegister classes have comments describing the next thing to implement. They need to handle the calls to create a new File of UserData for a new User, and to open and load a User's data File into the Model upon login. This means ReadFileData needs to be able to read and process UserData.