Added Model.login() and implementation of a MasterList of registered Usernames.
REMINDER: To run this in it's current form, run the individual file UserProfileDriver. DO NOT run the project.

Model.login() is untested.

Added JPasswordField for input password so password is disguised during registration.

Moved handling of invalid username and password during profile creation back into UserProfileModel.

Changed the File creation for a new registration to put the password in the first line, and the username in the second. This makes it easier to use the FileWrite() in FileData, which reads the first line of the file.