Read Description v2
Added FriendsList to ComboBox. I had to change HomePageView to take a String Array as a parameter to do this, so be careful.

With the way our GUI works, I don't think it is possible to have our ComboBox refresh whenever a new friend is added, so I just added a message that tells the User to log out and log back in.

Also, now whenever a new User is registered, it creates their Friends file in addition to their Data file.

Also finished Logout button on HomePageController.

THINGS TO DO:
- Whenever a friend is selected from the menu: an ActionListener in HomePageController reads their username and stores it in Model, Model initializes two conversation file names, checks if each file exists, and creates them if they don't. Use FileData to read the current User's conversation file into an ArrayList, and use a while loop to display it line-by-line in the JTextArea. Whenever a Message is sent, display it in the JTextArea, write it to the conversation file, and increment the User's level counter. (The level counter may not be properly implemented)

-Thoroughly comment and JAVADOC everything, and make sure the format is consistent

-If time, implement the User class for cleaner data storage